Initial commit of Climate Explorer CGI scripts into git
This is a snapshot of the Climate Explorer on 8 July 2013, as far as I know everything works.
It includes the bash ,.cgi scripts, the grads .gs scripts and one NCL .ncl file plus the HTML files and bin/*.sh
It also includes the U Reading R scripts that were put in in eth ENSEMBLES project in r/*.r
The Fortran files are maintained in a separate git repository.